#2 Login page updated with cute temp bg and logo and buttons
